+++ pennsieve/client.py
@@ -21,7 +21,7 @@ from .api.user import UserAPI
 from .base import ClientSession
 from .config import Settings
 from .models import Dataset, ModelTemplate
-
+from warnings import warn
 
 class Pennsieve(object):
     """
@@ -136,6 +136,7 @@ class Pennsieve(object):
         )
 
         self._api._context = self._api.organizations.get(self._api._organization)
+        warn(f'Pennsieve is transitioning to the new agent. This class \'{self.__class__.__name__}\' will be deprecated and API will significantly change; version=7.0.0; date=2022-11-01.', DeprecationWarning, stacklevel=2)
 
     @property
     def context(self):