+++ pennsieve/config.py
@@ -144,6 +144,7 @@ Additional environment variables and their corresponding config options:
 
 from __future__ import absolute_import, division, print_function
 
+from warnings import warn
 import configparser
 import os
 
@@ -238,6 +239,7 @@ class Settings(object):
         # check and create cache dir
         if not os.path.exists(self.cache_dir) and self.use_cache:
             os.makedirs(self.cache_dir)
+        warn(f'Pennsieve is transitioning to the new agent. This class \'{self.__class__.__name__}\' will be deprecated; version=7.0.0; date=2022-11-01.', DeprecationWarning, stacklevel=2)
 
     def _load_env(self):
         override = {}